+++ src/autocollapse/autocollapse.js
@@ -2,6 +2,7 @@
 
 import elementReady from 'element-ready';
 import ignore from 'ignore';
+import {toggleDiff} from '../collapse-diff/collapse-diff';
 
 export default {
     init,
@@ -29,9 +30,6 @@ async function collapseIfNeededAsync(section) {
         return;
     }
 
-    // Use the commented line below instead of the one below it when this issue #10 is resolved
-    // https://github.com/sindresorhus/element-ready/issues/10
-    // const button = await elementReady('.__refined_bitbucket_collapse_diff_button', {target: section});
-    const button = await elementReady(`section[data-identifier="${section.getAttribute('data-identifier')}"] .__refined_bitbucket_collapse_diff_button`);
-    button.click();
+    await elementReady('.__refined_bitbucket_collapse_diff_button', {target: section});
+    toggleDiff(section);
 }